users2fa/requirements/local.txt
asgiref==3.8.1
attrs==24.2.0
colorama==0.4.6
Django==5.1
django-cors-headers==4.4.0
django-model-utils==4.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
python-decouple==3.8
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Unipath==1.1
uritemplate==4.1.1
